+++ tml-sdk-spring-boot-starter-web/src/main/resources/log4j2.component.properties
@@ -1,0 +1,1 @@
+log4j2.threadContextMap=io.github.timemachinelab.log.mdc.TmlLogThreadContextMap
